+++ packages/react/src/components/Tree/types.ts
@@ -42,7 +42,6 @@ export interface TreeProps extends Omit<
   allowDeselect?: boolean
   loadData?: TreeLoadDataFn
   loadedKeys?: TreeNodeKey[]
-  filterValue?: string
   searchValue?: string
   defaultSearchValue?: string
   searchable?: boolean